step 2 - luo app.js
